Rename flexio1 to flexio for 1010, 1015, 1020 MCUs
Renaming the module to simply `flexio` allows consumers to use
consistent paths for all imxrt-ral MCUs. The renames are consistent with
similar renames like ADC and PWM.
